requirements.txt
flask
flask-cors
mysql-connector-python
gunicorn
python-dotenv
werkzeug
